+++ tests/test_visitors/test_ast/test_compares/conftest.py
@@ -2,11 +2,19 @@
 
 import pytest
 
+# Equality:
+
 if_with_is = 'if {0} is {1}: ...'
 if_with_is_not = 'if {0} is not {1}: ...'
 
 if_with_eq = 'if {0} == {1}: ...'
 if_with_not_eq = 'if {0} != {1}: ...'
+
+assert_construct = 'assert {0} == {1}'
+assert_with_message = 'assert {0} == {1}, "message"'
+
+# Not equality:
+
 if_with_gt = 'if {0} > {1}: ...'
 if_with_lt = 'if {0} < {1}: ...'
 if_with_gte = 'if {0} >= {1}: ...'
@@ -21,18 +29,22 @@ if_with_not_in = 'if {0} not in {1}: ...'
 
 ternary = 'ternary = 0 if {0} > {1} else 1'
 while_construct = 'while {0} > {1}: ...'
-assert_construct = 'assert {0} == {1}'
-assert_with_message = 'assert {0} == {1}, "message"'
 
 
 # Actual fixtures:
 
-@pytest.fixture(params=[
+EQUAL_COMPARES = frozenset((
     if_with_is,
     if_with_is_not,
 
     if_with_eq,
     if_with_not_eq,
+
+    assert_construct,
+    assert_with_message,
+))
+
+OTHER_COMPARES = frozenset((
     if_with_lt,
     if_with_gt,
     if_with_lte,
@@ -40,11 +52,24 @@ assert_with_message = 'assert {0} == {1}, "message"'
 
     ternary,
     while_construct,
-    assert_construct,
-    assert_with_message,
-])
+))
+
+
+@pytest.fixture(params=EQUAL_COMPARES | OTHER_COMPARES)
 def simple_conditions(request):
     """Fixture that returns simple conditionals."""
+    return request.param
+
+
+@pytest.fixture(params=EQUAL_COMPARES)
+def eq_conditions(request):
+    """Fixture that returns `eq` and `not eq` conditionals."""
+    return request.param
+
+
+@pytest.fixture(params=OTHER_COMPARES)
+def other_conditions(request):
+    """Fixture that returns `eq` and `not eq` conditionals."""
     return request.param